integration tests refactoring
