+++ .github/workflows/go.yml
@@ -110,7 +110,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        INTEGRATION_TEST_SUITE: [ "config","transfer" ]
+        INTEGRATION_TEST_SUITE: [ "config","emission" ]
     runs-on: 'ubuntu-20.04'
     steps:
       - uses: actions/checkout@v4
