+++ tests/integration/config/config_test.go
@@ -1,4 +1,4 @@
-package config_test
+package config
 
 import (
 	"github.com/anoideaopen/acl/tests/common"
